+++ payments/RELEASES.md
@@ -288,3 +288,13 @@ Based on:
 - [typescript v2.8.8] payments
 ### Releases
 - [NPM v2.8.8] https://www.npmjs.com/package/@wingspan/payments/v/2.8.8 - payments
+
+## 2025-01-13 01:32:21
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0-oas3.1 ./openapi/payment_speakeasy_edit.yaml
+- Speakeasy CLI 1.468.2 (2.493.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.8.9] payments
+### Releases
+- [NPM v2.8.9] https://www.npmjs.com/package/@wingspan/payments/v/2.8.9 - payments